DOMA-1138 move hooks folder outside components
